Adds content viewport layout
