+++ src/components/globalStyle.js
@@ -37,7 +37,6 @@ const GlobalStyle = createGlobalStyle`
     }
   }
   body {
-    margin: 0 auto;
     color: var(--main-text-color);
     font-family: -apple-system, BlinkMacSystemFont, Segoe UI, Roboto, Oxygen, Ubuntu, Cantarell, Fira Sans, Droid Sans, Helvetica Neue, sans-serif;
   }
