+++ src/pages/index.js
@@ -4,11 +4,13 @@ import Hero from '../components/sections/hero'
 import Work from '../components/sections/work'
 import Projects from '../components/sections/projects'
 import Contact from '../components/sections/contact'
+import About from '../components/sections/about'
 
 const IndexPage = () => {
   return (
     <Layout>
       <Hero />
+      <About />
       <Projects />
       <Work />
       <Contact />
